+++ awc/CHANGES.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+- Add `hickory-resolver` crate feature (off-by-default).
+- The `trust-dns` crate feature now delegates DNS resolution to `hickory-dns`.
+
 ## 3.7.0
 
 - Update `brotli` dependency to `8`.
@@ -12,7 +15,6 @@
 - Do not send `Host` header on HTTP/2 requests, as it is not required, and some web servers may reject it.
 - Update `brotli` dependency to `7`.
 - Minimum supported Rust version (MSRV) is now 1.75.
-- Replace `trust-dns-resolver` with `hickory-resolver` for DNS resolution in awc when using `trust-dns` feature.
 
 ## 3.5.1